fix: フロントエンドの型をi32に統一し、注意書きコメントを追加
- src/types/gamification.rs: LevelInfo, XpGainedEvent, SyncResultの型をu32からi32に変更
- バックエンド・フロントエンド両方に型選択の理由を説明するコメントを追加
- src/components/home/mod.rs: 不要なas u32キャストを削除

Note: XPやレベルは意味的にはu32が適切だが、SQLiteのINTEGER型が符号あり整数であり、
sqlxがi32としてマッピングするため、DB層との整合性を保つためにi32で統一

+++ src/components/home/mod.rs
@@ -53,7 +53,7 @@ fn handle_sync_result_notifications(
         if should_show_app_notification {
             let event = XpGainedEvent {
                 xp_gained: sync_result.xp_gained,
-                total_xp: sync_result.user_stats.total_xp as u32,
+                total_xp: sync_result.user_stats.total_xp,
                 old_level: sync_result.old_level,
                 new_level: sync_result.new_level,
                 level_up: sync_result.level_up,
